requirements.txt
Flask
redis-py
requests
